+++ src/__mocks__/accountNotify.ts
@@ -4,7 +4,7 @@ const accountChanged = {
   cmd: 'accountChanged',
 };
 
-export const accountChangedHeaders = [
+export const accountChangedHeaders: [string, string][] = [
   [HEADER_CONTENT_TYPE_NAME, CONTENT_TYPE_NOTIFY],
   [HEADER_NOTIFY, JSON.stringify(accountChanged)],
 ];
@@ -13,7 +13,7 @@ const accountDeleted = {
   cmd: 'accountDeleted',
 };
 
-export const accountDeletedHeaders = [
+export const accountDeletedHeaders: [string, string][] = [
   [HEADER_CONTENT_TYPE_NAME, CONTENT_TYPE_NOTIFY],
   [HEADER_NOTIFY, JSON.stringify(accountDeleted)],
 ];