+++ src/__mocks__/channels.ts
@@ -10,7 +10,7 @@ const outputChannels = '2,3';
 const headerInputChannels = `${HEADER_INPUT_CHANNELS}: ${inputChannels}`;
 const headerOutputChannels = `${HEADER_OUTPUT_CHANNELS}: ${outputChannels}`;
 
-export const channelsHeaders = [
+export const channelsHeaders: [string, string][] = [
   [HEADER_CONTENT_TYPE_NAME, CONTENT_TYPE_ENTER_ROOM],
   [HEADER_INPUT_CHANNELS, inputChannels],
   [HEADER_OUTPUT_CHANNELS, outputChannels],