+++ src/__mocks__/channelsNotify.ts
@@ -3,7 +3,7 @@ import { HEADER_CONTENT_TYPE_NAME, CONTENT_TYPE_NOTIFY, HEADER_NOTIFY } from '..
 const channelsStringified = JSON.stringify({ cmd: 'channels', input: '0,1', output: '0' });
 const channelsParsed = JSON.parse(channelsStringified);
 
-export const channelsHeaders = [
+export const channelsHeaders: [string, string][] = [
   [HEADER_CONTENT_TYPE_NAME, CONTENT_TYPE_NOTIFY],
   [HEADER_NOTIFY, channelsStringified],
 ];