+++ CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [6.3.3](https://github.com/Krivega/sip-connector/compare/v6.3.2...v6.3.3) (2022-08-28)
+
+### [6.3.2](https://github.com/Krivega/sip-connector/compare/v6.3.1...v6.3.2) (2022-08-12)
+
+### Bug Fixes
+
+- calls without mics or cams ([113d8ff](https://github.com/Krivega/sip-connector/commit/113d8ff18ee2be318fc11aa0f78fdea361b38836))
+
+### [6.3.1](https://github.com/Krivega/sip-connector/compare/v6.3.0...v6.3.1) (2022-08-08)
+
+## [6.3.0](https://github.com/Krivega/sip-connector/compare/v6.2.0...v6.3.0) (2022-08-08)
+
+### Features
+
+- add videoMode, audioMode options to answer ([30ca8e5](https://github.com/Krivega/sip-connector/commit/30ca8e50a49fb284de83af499e5d60e15dcaeb59))
+
+## [6.2.0](https://github.com/Krivega/sip-connector/compare/v6.1.0...v6.2.0) (2022-08-05)
+
+### Features
+
+- add videoMode, audioMode options to call ([0b9d3db](https://github.com/Krivega/sip-connector/commit/0b9d3db2ad92641f508a83e4d4945c172e7d3a29))
+
+## [6.1.0](https://github.com/Krivega/sip-connector/compare/v6.0.2...v6.1.0) (2022-07-11)
+
+### Features
+
+- move userAgent to params of connect ([#23](https://github.com/Krivega/sip-connector/issues/23)) ([12093af](https://github.com/Krivega/sip-connector/commit/12093af714f896cfed91596875146baf64236efd))
+
+### [6.0.2](https://github.com/Krivega/sip-connector/compare/v6.0.1...v6.0.2) (2022-06-07)
+
 ### [6.0.1](https://github.com/Krivega/sip-connector/compare/v6.0.0...v6.0.1) (2022-05-31)
 
 ### Bug Fixes